refactor: retry 로직 분리
